Replace old getters with new ones.
