+++ src/commands/disable.js
@@ -1,5 +1,5 @@
 const actions = require("../store/actions");
-const { getters } = require("../store/getters");
+const getters = require("../store/getters");
 const telegramUtility = require("../util/telegram");
 
 /**
@@ -8,7 +8,7 @@ const telegramUtility = require("../util/telegram");
 const disable = ({ store }) => ctx => {
   const chatId = telegramUtility.chatIdInContext(ctx);
 
-  if (getters.isChatActive(chatId, store)) {
+  if (getters.isChatEnabled(chatId)(store.getState())) {
     store.dispatch(actions.disableChat(chatId));
     ctx.reply(ctx.t("command.disable.disabled"));
   } else {
